Mem Leak fixed + cleanups
