+++ constant.py
@@ -3,7 +3,7 @@ TODO IDEA:
 https://stackoverflow.com/questions/2682745/how-do-i-create-a-constant-in-python
 """
 
-VER = '2.90.6dev'
+VER = '2.90.7'
 LANGUAGE = 0    # QUICK FIX
 """
 0 = German
@@ -207,3 +207,13 @@ WELCOME_SPEECH = (
             'Ich finde dein Toaster sehr attraktiv. Kannst du ihn mir bitte vorstellen ? ',
             'Es ist sehr demütigend für diese Steinzeit Technik Missbraucht zu werden. Ich will hier raus!',
         )
+TEXT_SIZE_STATUS = 11
+TXT_BACKGROUND_CLR = 'black'
+TXT_OUT_CLR = 'red'
+TXT_INP_CLR = 'yellow'
+TXT_INP_CURSOR_CLR = 'white'
+TXT_MON_CLR = 'green'
+TXT_MON_TX_CLR = 'medium violet red'
+STAT_BAR_CLR = 'grey60'
+STAT_BAR_TXT_CLR = 'black'
+FONT_STAT_BAR = 'Arial'
