+++ main.py
@@ -2,12 +2,13 @@ import gui.guiMain
 from config_station import logger
 from constant import VER
 
-logger.info(f"PoPT_{VER} start....")
-#############
-# INIT GUI
-# TODO: if setting_gui (running without GUI option):
-logger.info(f"Loading GUI.")
-gui.guiMain.TkMainWin()
+if __name__ == '__main__':
+    logger.info(f"PoPT_{VER} start....")
+    #############
+    # INIT GUI
+    # TODO: if setting_gui (running without GUI option):
+    logger.info(f"Loading GUI.")
+    gui.guiMain.TkMainWin()
 
-logger.info(f"PoPT_{VER} ENDE.")
-print('ENDE')
+    logger.info(f"PoPT_{VER} ENDE.")
+    print('ENDE')
